implemented templates in vector class

+++ OOGL/include/OOGL/ooglTime.hpp
@@ -3,7 +3,7 @@
 
 namespace oogl
 {
-	double getTime();
+	int getTime();
 }
 
 #endif
